# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Brief": {
        "description": "El encargo de la obra.\n\n`start` es el instante en que arranca el mundo: todo lo que el brief\nestablece es cierto desde ahi, y nada anterior existe.\n\nLo que va despues de `word_tolerance` lo trae la entrevista (RF-200). Es\nopcional para que un brief escrito a mano, sin destinatario, siga valiendo.",
        "properties": {
          "dedication": {
            "default": "",
            "title": "Dedication",
            "type": "string"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/BriefEntity"
            },
            "title": "Entities",
            "type": "array"
          },
          "forbidden_themes": {
            "items": {
              "type": "string"
            },
            "title": "Forbidden Themes",
            "type": "array"
          },
          "forbidden_words": {
            "items": {
              "type": "string"
            },
            "title": "Forbidden Words",
            "type": "array"
          },
          "genre": {
            "default": "",
            "title": "Genre",
            "type": "string"
          },
          "recipient": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recipient"
              },
              {
                "type": "null"
              }
            ]
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/BriefRelation"
            },
            "title": "Relations",
            "type": "array"
          },
          "rulebook": {
            "default": "",
            "description": "DEP-02, vacio si no es deportiva",
            "title": "Rulebook",
            "type": "string"
          },
          "start": {
            "$ref": "#/components/schemas/WorldTime"
          },
          "style_guide": {
            "description": "POE-06",
            "minLength": 1,
            "title": "Style Guide",
            "type": "string"
          },
          "target_words": {
            "exclusiveMinimum": 0.0,
            "title": "Target Words",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "tone": {
            "default": "",
            "title": "Tone",
            "type": "string"
          },
          "word_tolerance": {
            "default": 0.1,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Word Tolerance",
            "type": "number"
          }
        },
        "required": [
          "title",
          "start",
          "entities",
          "style_guide",
          "target_words"
        ],
        "title": "Brief",
        "type": "object"
      },
...
      "NovelCreated": {
        "properties": {
          "events_loaded": {
            "description": "Hechos que el brief aporto al canon",
            "title": "Events Loaded",
            "type": "integer"
          },
          "novel_id": {
            "title": "Novel Id",
            "type": "string"
          }
        },
        "required": [
          "novel_id",
          "events_loaded"
        ],
        "title": "NovelCreated",
        "type": "object"
      },
...
      "NovelList": {
        "properties": {
          "novels": {
            "items": {
              "$ref": "#/components/schemas/NovelEntry"
            },
            "title": "Novels",
            "type": "array"
          }
        },
        "required": [
          "novels"
        ],
        "title": "NovelList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Sistema autonomo de generacion de novelas largas. La API observa y arranca; no aprueba, no corrige y no desbloquea.",
    "title": "Story Maker",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/novels": {
      "get": {
        "description": "RI-37. Las novelas del directorio de tiradas, con su estado y sus versiones.\n\nVive aqui y no en `canon/` porque el estado de la tirada es del Orquestador.",
        "operationId": "list_novels_novels_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Novels",
        "tags": [
          "orchestration"
        ]
      },
      "post": {
        "description": "RI-01. Carga el brief. Es la unica escritura que la API provoca.",
        "operationId": "create_novels_post",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "query",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Brief"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "canon"
        ]
      }
    },
...
    "/novels/{novel_id}/chapters": {
      "get": {
        "description": "RI-04. Solo capitulos congelados: son los unicos que estan en el indice.",
        "operationId": "list_chapters_novels__novel_id__chapters_get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Chapters",
        "tags": [
          "canon"
        ]
      }
    },
    "/novels/{novel_id}/chapters/{number}": {
      "get": {
        "description": "RI-05. La prosa de un capitulo congelado.",
        "operationId": "get_chapter_novels__novel_id__chapters__number__get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "title": "Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterProse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chapter",
        "tags": [
          "canon"
        ]
      }
    },
...
    "/novels/{novel_id}/entities": {
      "get": {
        "description": "RI-45. Personajes, lugares, instituciones y objetos, con los capitulos donde aparecen.",
        "operationId": "list_entities_novels__novel_id__entities_get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          },
          {
            "description": "PER-01, MUN-01, MUN-03 o MUN-02",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "person",
                    "place",
                    "institution",
                    "object"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "PER-01, MUN-01, MUN-03 o MUN-02",
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Entities",
        "tags": [
          "canon"
        ]
      }
    },
    "/novels/{novel_id}/entities/{entity_id}": {
      "get": {
        "description": "RI-46. La ficha: compacta, hechos con procedencia, relaciones vigentes y apariciones.",
        "operationId": "get_entity_novels__novel_id__entities__entity_id__get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityFile"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Entity",
        "tags": [
          "canon"
        ]
      }
    },
...
    "/novels/{novel_id}/retcons": {
      "get": {
        "description": "RI-31. Los retcons aplicados, con la regla que los admitio.",
        "operationId": "list_retcons_novels__novel_id__retcons_get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetconList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Retcons",
        "tags": [
          "canon"
        ]
      }
    },
...
    "/novels/{novel_id}/state": {
      "get": {
        "description": "RI-06. El estado del mundo en un instante.\n\nEl instante es obligatorio: un estado sin instante no significa nada en un\nsistema donde todo tiene vigencia.\n\nEl formato se declara **en el contrato** y no solo en el validador. Mientras\nvivio solo en el codigo, el esquema prometia aceptar cualquier texto y la\nruta rechazaba casi todos: un cliente generado desde ese contrato habria\nenviado basura convencido de que era valida. Lo encontro la comprobacion\ngenerativa, que es para lo que esta.",
        "operationId": "get_state_novels__novel_id__state_get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          },
          {
            "description": "Instante de mundo, ISO 8601",
            "in": "query",
            "name": "at",
            "required": true,
            "schema": {
              "description": "Instante de mundo, ISO 8601",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}(T\\d{2}:\\d{2}(:\\d{2})?)?$",
              "title": "At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Seq",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldState"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get State",
        "tags": [
          "canon"
        ]
      }
    },
...
    "/novels/{novel_id}/versions": {
      "get": {
        "description": "RI-42. Versiones del manuscrito (PRO-08): la 1 es la tirada, cada enmienda una mas.",
        "operationId": "list_versions_novels__novel_id__versions_get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Versions",
        "tags": [
          "canon"
        ]
      }
    },
...
  }
}